typescript(vx-demo): rewrite BarStackHorizontal in TypeScript
