+++ packages/vx-demo/src/components/tiles/BarStackHorizontal.tsx
@@ -1,0 +1,204 @@
+import React from 'react';
+import { BarStackHorizontal } from '@vx/shape';
+import { Group } from '@vx/group';
+import { AxisBottom, AxisLeft } from '@vx/axis';
+import cityTemperature, { CityTemperature } from '@vx/mock-data/lib/mocks/cityTemperature';
+import { scaleBand, scaleLinear, scaleOrdinal } from '@vx/scale';
+import { timeParse, timeFormat } from 'd3-time-format';
+import { withTooltip, Tooltip } from '@vx/tooltip';
+import { WithTooltipProvidedProps } from '@vx/tooltip/lib/enhancers/withTooltip';
+import { LegendOrdinal } from '@vx/legend';
+import { ShowProvidedProps } from '../../types';
+
+type CityName = 'New York' | 'San Francisco' | 'Austin';
+type TooltipData = {
+  bar: {
+    0: number;
+    1: number;
+    data: CityTemperature;
+  };
+  key: CityName;
+  index: number;
+  height: number;
+  width: number;
+  x: number;
+  y: number;
+  color: string;
+};
+
+const purple1 = '#6c5efb';
+const purple2 = '#c998ff';
+const purple3 = '#a44afe';
+const bg = '#eaedff';
+
+const data = cityTemperature.slice(0, 12);
+const keys = Object.keys(data[0]).filter(d => d !== 'date') as CityName[];
+
+const temperatureTotals = data.reduce((allTotals, currentDate) => {
+  const totalTemperature = keys.reduce((dailyTotal, k) => {
+    dailyTotal += Number(currentDate[k]);
+    return dailyTotal;
+  }, 0);
+  allTotals.push(totalTemperature);
+  return allTotals;
+}, [] as number[]);
+
+const parseDate = timeParse('%Y%m%d');
+const format = timeFormat('%b %d');
+const formatDate = (date: string) => format(parseDate(date) as Date);
+
+// accessors
+const getDate = (d: CityTemperature) => d.date;
+
+// scales
+const temperatureScale = scaleLinear<number>({
+  domain: [0, Math.max(...temperatureTotals)],
+  nice: true,
+});
+const dateScale = scaleBand<string>({
+  domain: data.map(getDate),
+  padding: 0.2,
+});
+const colorScale = scaleOrdinal<CityName, string>({
+  domain: keys,
+  range: [purple1, purple2, purple3],
+});
+
+let tooltipTimeout: number;
+
+export default withTooltip<ShowProvidedProps, TooltipData>(
+  ({
+    width,
+    height,
+    events = false,
+    margin = {
+      top: 40,
+      left: 50,
+      right: 40,
+      bottom: 100,
+    },
+    tooltipOpen,
+    tooltipLeft,
+    tooltipTop,
+    tooltipData,
+    hideTooltip,
+    showTooltip,
+  }: ShowProvidedProps & WithTooltipProvidedProps<TooltipData>) => {
+    if (width < 10) return null;
+
+    // bounds
+    const xMax = width - margin.left - margin.right;
+    const yMax = height - margin.top - margin.bottom;
+
+    temperatureScale.rangeRound([0, xMax]);
+    dateScale.rangeRound([yMax, 0]);
+
+    return (
+      <div style={{ position: 'relative' }}>
+        <svg width={width} height={height}>
+          <rect width={width} height={height} fill={bg} rx={14} />
+          <Group top={margin.top} left={margin.left}>
+            <BarStackHorizontal<CityTemperature, CityName>
+              data={data}
+              keys={keys}
+              height={yMax}
+              y={getDate}
+              xScale={temperatureScale}
+              yScale={dateScale}
+              color={colorScale}
+            >
+              {barStacks =>
+                barStacks.map(barStack =>
+                  barStack.bars.map(bar => (
+                    <rect
+                      key={`barstack-horizontal-${barStack.index}-${bar.index}`}
+                      x={bar.x}
+                      y={bar.y}
+                      width={bar.width}
+                      height={bar.height}
+                      fill={bar.color}
+                      onClick={() => {
+                        if (events) alert(`clicked: ${JSON.stringify(bar)}`);
+                      }}
+                      onMouseLeave={() => {
+                        tooltipTimeout = window.setTimeout(() => {
+                          hideTooltip();
+                        }, 300);
+                      }}
+                      onMouseMove={() => {
+                        if (tooltipTimeout) clearTimeout(tooltipTimeout);
+                        const top = bar.y + margin.top;
+                        const left = bar.x + bar.width + margin.left;
+                        showTooltip({
+                          tooltipData: bar,
+                          tooltipTop: top,
+                          tooltipLeft: left,
+                        });
+                      }}
+                    />
+                  )),
+                )
+              }
+            </BarStackHorizontal>
+            <AxisLeft<string>
+              hideAxisLine
+              hideTicks
+              scale={dateScale}
+              tickFormat={formatDate}
+              stroke={purple3}
+              tickStroke={purple3}
+              tickLabelProps={() => ({
+                fill: purple3,
+                fontSize: 11,
+                textAnchor: 'end',
+                dy: '0.33em',
+              })}
+            />
+            <AxisBottom<number>
+              top={yMax}
+              scale={temperatureScale}
+              stroke={purple3}
+              tickStroke={purple3}
+              tickLabelProps={() => ({
+                fill: purple3,
+                fontSize: 11,
+                textAnchor: 'middle',
+              })}
+            />
+          </Group>
+        </svg>
+        <div
+          style={{
+            position: 'absolute',
+            top: margin.top / 2 - 10,
+            width: '100%',
+            display: 'flex',
+            justifyContent: 'center',
+            fontSize: '14px',
+          }}
+        >
+          <LegendOrdinal scale={colorScale} direction="row" labelMargin="0 15px 0 0" />
+        </div>
+        {tooltipOpen && tooltipData && (
+          <Tooltip
+            top={tooltipTop}
+            left={tooltipLeft}
+            style={{
+              minWidth: 60,
+              backgroundColor: 'rgba(0,0,0,0.9)',
+              color: 'white',
+            }}
+          >
+            <div style={{ color: colorScale(tooltipData.key) }}>
+              <strong>{tooltipData.key}</strong>
+            </div>
+            <div>{tooltipData.bar.data[tooltipData.key]}℉</div>
+            <div>
+              <small>{formatDate(getDate(tooltipData.bar.data))}</small>
+            </div>
+          </Tooltip>
+        )}
+      </div>
+    );
+  },
+);
